Use root time for self time for inverted stacks
Fixes #389

This should cover what we need. The trick with this was keeping
total time the same as before while changing self time to be on
the roots (trunks, first branches?) of the tree. In order to do
so I had to separate "self" time, which is a UI value, from
"leaf" time, which we need to use to calculate the total time
of parents.